Moved getOidcConfig to OAuth2.ts
