+++ src/pages/Main.ts
@@ -16,7 +16,7 @@ import FirecloudNotification from 'src/components/FirecloudNotification';
 import IdleStatusMonitor from 'src/components/IdleStatusMonitor';
 import SupportRequest from 'src/components/SupportRequest';
 import { TitleManager } from 'src/components/TitleManager';
-import { getOidcConfig } from 'src/libs/auth';
+import { getOidcConfig } from 'src/libs/ajax/OAuth2';
 import { getEnabledBrand } from 'src/libs/brand-utils';
 import { reportError } from 'src/libs/error';
 import { PageViewReporter } from 'src/libs/events';
